// src/app/models/user.model.ts
export interface UserResponse {
  id: number;
  name: string;
  username: string;
  email: string;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}

export interface Geo {
  lat: string;
  lng: string;
}